+++ src/bind/cmdbind.c
@@ -29,6 +29,7 @@ SCUT(save)
 SCUT(goto)
 SCUT(quit)
 SCUT(saveall)
+SCUT(buffer)
 
 RCUT(swap)
 RCUT(snap)
@@ -50,6 +51,7 @@ void bind_cmdbind_init(void)
     SCUT_ADD(goto);
     SCUT_ADD(quit);
     SCUT_ADD(saveall);
+    SCUT_ADD(buffer);
 
     RCUT_ADD(swap);
     RCUT_ADD(snap);